requirements.txt
# Smart Document Assistant with OCR + NLP  —  requirements.txt
# Python 3.9+  |  Install: pip install -r requirements.txt

# NLP / Transformers
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

# PDF Extraction
PyMuPDF>=1.23.0

# OCR backends
pytesseract>=0.3.10
# System binary also needed:
#   Ubuntu:  sudo apt install tesseract-ocr tesseract-ocr-urd
#   macOS:   brew install tesseract
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
easyocr>=1.7.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# UI
streamlit>=1.28.0

# Utilities
regex>=2023.0.0

# Testing
pytest>=7.4.0
